app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose
python-dotenv
email-validator
passlib[bcrypt]==1.7.4
bcrypt<4.1.0
python-multipart